requirements.txt
python_dotenv==0.15
discord.py==1.5
Pillow==8.0
